Mass dump of arm subsystem updates
- Add MoveToIntakeCommand.
- Add boundary classes.
- Add no-go zones.
- Define no-go zones.
- Add safety points.
- Add reach points.

- Emmie Jones

+++ src/main/include/Constants.h
@@ -5,11 +5,11 @@ namespace Constants {
     const int k_NumberOfSwerveModules = 4;
 
     namespace Arm {
-        const float k_forearmLenInches = 24.0; // Update to correct value
-        const float k_bicepLenInches = 24.0;// Update to correct value
+        const double k_forearmLenInches = 24.0; // Update to correct value
+        const double k_bicepLenInches = 24.0;// Update to correct value
 
-        const int k_shoulderMinDeg = 0;
-        const int k_shoulderMaxDeg = 270;
+        const double k_shoulderMinDeg = 0;
+        const double k_shoulderMaxDeg = 270;
 
     };
 };
